Small updates, change to syntax_check()
syntax_check() now returns an integer instead of a boolean.  This allows
error codes to be easily compared for use in a testing program.  A key
to the error codes exists in the file ERROR CODE KEY.txt.

I added a default case to the syntax_check() switch statement, as well
as improving the error messages when checking the syntax.

These changes have not been tested yet as my copy of VS2013 on my laptop
is being dumb.