Still need error help ASAP
Here is code that should work for booleans and arithmetic, still needs testing but for now we need to solve linking errors.  I am getting unresolved external symbol errors in this code and I cannot find out why.  After that all this needs is some quick testing til it's polished and works!